Avoid passing handler as an argument to python
The transpiler can now be called directly from the command line instead
of passing as an argument to python. In the previous version, the user
had to type "python _handler.py test.ty" to transpile and run the file
test.ty. Now, the user can simply type "./transpiler test.ty" for
the same.