Utilize Python's built-in tokenizer
In this commit, Python's built-in tokenizer library is utilized to detect
strings and comments instead of regular expressions. Since the library
is designed to parse syntactically valid Python code, it doesn't work on
non-Python source code. However, as the transpy code is syntactically
similar to Python, the library can be used to scan strings and comments.